requirements.txt
strawberry-graphql[debug-server]
uvicorn[standard]
strawberry-graphql
fastapi